User Types

+++ BountyfulSushi/Controllers/BountyController.cs
@@ -7,6 +7,7 @@ using System.Security.Claims;
 using System.Collections.Generic;
 using BountyfulSushi.Repositories.Interfaces;
 using BountyfulSushi.Models;
+using BountyfulSushi.Repositories;
 
 namespace BountyfulSushi.Controllers
 {
@@ -15,9 +16,11 @@ namespace BountyfulSushi.Controllers
     [Authorize]
     public class BountyController : ControllerBase
     {
+        private readonly IUserRepository _userRepository;
         private readonly IBountyRepository _bountyRepository;
-        public BountyController(IBountyRepository bountyRepository)
+        public BountyController(IBountyRepository bountyRepository, IUserRepository userRepository)
         {
+            _userRepository = userRepository;
             _bountyRepository = bountyRepository;
         }
 
@@ -45,6 +48,13 @@ namespace BountyfulSushi.Controllers
         [HttpPost]
         public IActionResult Post(Bounty bounty)
         {
+            var currentUser = GetCurrentUser();
+
+            if (currentUser.UserType.Id != 1)
+            {
+                return Unauthorized();
+            }
+
             _bountyRepository.Add(bounty);
             return CreatedAtAction("Get", new { id = bounty.Id }, bounty);
         }
@@ -53,6 +63,13 @@ namespace BountyfulSushi.Controllers
         [HttpPut("{id}")]
         public IActionResult Put(Bounty bounty)
         {
+            var currentUser = GetCurrentUser();
+
+            if (currentUser.UserType.Id != 1)
+            {
+                return Unauthorized();
+            }
+
             _bountyRepository.Update(bounty);
             return NoContent();
         }
@@ -61,8 +78,21 @@ namespace BountyfulSushi.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
+            var currentUser = GetCurrentUser();
+
+            if (currentUser.UserType.Id != 1)
+            {
+                return Unauthorized();
+            }
+
             _bountyRepository.Delete(id);
             return NoContent();
+        }
+
+        private User GetCurrentUser()
+        {
+            var fireBaseId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            return _userRepository.GetByFireBaseId(fireBaseId);
         }
     }
 }
